[WIP] drivers: i2c: stm32: Add bus recovery in RTIO drivers
Add bus recovery in STM32 I2C RTIO drivers using the common sequence
that is factorized and used for both non-RTIO and RTIO drivers
implementation.

The async start sequence change needed to handle recovery requests
was dumped from https://github.com/zephyrproject-rtos/zephyr/pull/105712.
Patch still under review and suject to discussion.
