+++ drivers/i2c/Kconfig.stm32
@@ -47,7 +47,7 @@ config I2C_STM32_COMBINED_INTERRUPT
 
 config I2C_STM32_BUS_RECOVERY
 	bool "Bus recovery support"
-	default y if I2C_BUS_RECOVERY
+	default y if I2C_BUS_RECOVERY && GPIO
 	select I2C_BITBANG
 	help
 	  Enable STM32 driver bus recovery support via GPIO bitbanging.
