+++ drivers/i2c/i2c_stm32_common.c
@@ -11,6 +11,11 @@
 #include <zephyr/drivers/clock_control.h>
 #include <zephyr/drivers/pinctrl.h>
 
+#ifdef CONFIG_I2C_STM32_BUS_RECOVERY
+#include "i2c_bitbang.h"
+#include "i2c-priv.h"
+#endif /* CONFIG_I2C_STM32_BUS_RECOVERY */
+
 #define LOG_LEVEL CONFIG_I2C_LOG_LEVEL
 #include <zephyr/logging/log.h>
 LOG_MODULE_REGISTER(i2c_ll_stm32_common);
@@ -110,3 +115,104 @@ int i2c_stm32_pm_action(const struct device *dev, enum pm_device_action action)
 	return err;
 }
 #endif
+
+#if CONFIG_I2C_STM32_BUS_RECOVERY
+static void i2c_stm32_bitbang_set_scl(void *io_context, int state)
+{
+	const struct i2c_stm32_config *config = io_context;
+
+	gpio_pin_set_dt(&config->scl, state);
+}
+
+static void i2c_stm32_bitbang_set_sda(void *io_context, int state)
+{
+	const struct i2c_stm32_config *config = io_context;
+
+	gpio_pin_set_dt(&config->sda, state);
+}
+
+static int i2c_stm32_bitbang_get_sda(void *io_context)
+{
+	const struct i2c_stm32_config *config = io_context;
+
+	return gpio_pin_get_dt(&config->sda) == 0 ? 0 : 1;
+}
+
+int i2c_stm32_recover_bus(const struct device *dev)
+{
+	__maybe_unused struct i2c_stm32_data *data = dev->data;
+	const struct i2c_stm32_config *config = dev->config;
+	struct i2c_bitbang bitbang_ctx;
+	struct i2c_bitbang_io bitbang_io = {
+		.set_scl = i2c_stm32_bitbang_set_scl,
+		.set_sda = i2c_stm32_bitbang_set_sda,
+		.get_sda = i2c_stm32_bitbang_get_sda,
+	};
+	uint32_t bitrate_cfg = i2c_map_dt_bitrate(config->bitrate) | I2C_MODE_CONTROLLER;
+	int error = 0;
+
+	LOG_ERR("attempting to recover bus");
+
+	if (config->scl.port == NULL || config->sda.port == NULL) {
+		LOG_ERR("No SCL or SDA GPIO deefined");
+		return -ENOSYS;
+	}
+
+	if (!gpio_is_ready_dt(&config->scl)) {
+		LOG_ERR("SCL GPIO device not ready");
+		return -EIO;
+	}
+
+	if (!gpio_is_ready_dt(&config->sda)) {
+		LOG_ERR("SDA GPIO device not ready");
+		return -EIO;
+	}
+
+#ifndef CONFIG_I2C_RTIO
+	k_sem_take(&data->bus_mutex, K_FOREVER);
+#endif
+
+	error = gpio_pin_configure_dt(&config->scl, GPIO_OUTPUT_HIGH);
+	if (error != 0) {
+		LOG_ERR("failed to configure SCL GPIO (err %d)", error);
+		goto restore;
+	}
+
+	error = gpio_pin_configure_dt(&config->sda, GPIO_OUTPUT_HIGH);
+	if (error != 0) {
+		LOG_ERR("failed to configure SDA GPIO (err %d)", error);
+		goto restore;
+	}
+
+	i2c_bitbang_init(&bitbang_ctx, &bitbang_io, (void *)config);
+
+	error = i2c_bitbang_configure(&bitbang_ctx, bitrate_cfg);
+	if (error != 0) {
+		LOG_ERR("failed to configure I2C bitbang (err %d)", error);
+		goto restore;
+	}
+
+	error = i2c_bitbang_recover_bus(&bitbang_ctx);
+	if (error != 0) {
+		LOG_ERR("failed to recover bus (err %d)", error);
+	}
+
+restore:
+	(void)pinctrl_apply_state(config->pcfg, PINCTRL_STATE_DEFAULT);
+
+	/* Re-initialize the I2C peripheral after GPIO-based bus recovery.
+	 * pinctrl_apply_state() restores the pin configuration, but the
+	 * peripheral registers remain in a faulted state. Re-running
+	 * runtime_configure() restores the peripheral to a working state.
+	 */
+	if (i2c_stm32_runtime_configure(dev, bitrate_cfg) != 0) {
+		LOG_ERR("failed to restore I2C peripheral after bus recovery");
+	}
+
+#ifndef CONFIG_I2C_RTIO
+	k_sem_give(&data->bus_mutex);
+#endif
+
+	return error;
+}
+#endif /* CONFIG_I2C_STM32_BUS_RECOVERY */
